feat(lns-cli)!: split `lns artifact` from `lns sandbox`
§1.1: two things carry a name. An artifact is one `lns.run/v1` document,
published as one OCI artifact; a sandbox is what a sandbox artifact becomes
when it runs. One namespace each.

`lns artifact` takes init, validate, push, pull, tag, ls, inspect, rm and
prune. `lns sandbox` keeps run, exec, kill, stop, logs, attach, ls, inspect
and rm. `lns sandbox ps` is gone — `ls` is the verb and `lns ps` its shortcut
(§3.2) — and so are `lns sandbox init/validate/push/pull/tag/prune`.

`lns rm` and `lns inspect` are the two verbs both namespaces answer for, so
§1.3 arbitration lands with them: a path-shaped operand is the document and
the service is never asked; otherwise both namespaces are asked, one match
runs, two matches refuse by printing the two namespaced commands, and no
match names both as searched. `lns` never guesses. `prune` has no shortcut
because it takes no operand — nothing could tell the two apart.

The module tree follows the surface. `author`, `distribute`, `fileset` and
`mixin_plan` move under `artifact/`, along with the verbs that drive them and
their tests. `SandboxService` and `TermInfo` move to `service::client`, and
`RealSandboxService` to `service::real`, so neither namespace depends on the
other; `CannedService` moves to a shared `test_service` for the same reason.
`format_bytes` moves to `output`, where every table renders through.
