+++ crates/e2e-tests/features/sandbox_authoring.feature
@@ -3,7 +3,7 @@ Feature: authoring a sandbox definition offline
   working directory — no daemon, no network. These scenarios drive the
   real binary against a real project directory: init scaffolds the
   definition, validate judges it offline, and a target-less inspect renders the effective
-  definition. `lns init` is the top-level shortcut for `lns sandbox init`.
+  definition. `lns init` is the top-level shortcut for `lns artifact init`.
 
   Scenario: init scaffolds a default lns.yaml
     When I run "lns init" in the project directory
@@ -16,19 +16,19 @@ Feature: authoring a sandbox definition offline
 
   Scenario: validate refuses a bare tool name through the real binary
     Given a project definition declaring tool "node"
-    When I run "lns sandbox validate" in the project directory
+    When I run "lns artifact validate" in the project directory
     Then the exit code is non-zero
     And the output contains "node@latest"
 
   Scenario: validate refuses an engine backend prefix through the real binary
     Given a project definition declaring tool "npm:some-tool@3"
-    When I run "lns sandbox validate" in the project directory
+    When I run "lns artifact validate" in the project directory
     Then the exit code is non-zero
     And the output contains "engine backend prefix"
 
   Scenario: inspect lists the declared tools offline
     Given a project definition declaring tool "node@22"
-    When I run "lns sandbox inspect" in the project directory
+    When I run "lns artifact inspect" in the project directory
     Then the exit code is 0
     And the output contains "tool: node@22"
 
@@ -39,19 +39,19 @@ Feature: authoring a sandbox definition offline
     And the output contains "already exists"
 
   Scenario: validate passes the scaffolded definition
-    When I run "lns sandbox init" in the project directory
-    And I run "lns sandbox validate" in the project directory
+    When I run "lns artifact init" in the project directory
+    And I run "lns artifact validate" in the project directory
     Then the exit code is 0
     And the output contains "lns.yaml is valid."
 
   Scenario: validate without a definition points at lns init
-    When I run "lns sandbox validate" in the project directory
+    When I run "lns artifact validate" in the project directory
     Then the exit code is non-zero
     And the output contains "run `lns init` to scaffold one"
 
   Scenario: inspect with no target renders the effective definition
     When I run "lns init" in the project directory
-    And I run "lns sandbox inspect" in the project directory
+    And I run "lns artifact inspect" in the project directory
     Then the exit code is 0
     And the output contains "Sandbox: sandbox"
     And the output contains "docker.io/library/alpine:3.20"
