+++ crates/e2e-tests/features/sandbox_management.feature
@@ -1,5 +1,5 @@
 Feature: cached sandbox management end to end
-  `lns sandbox` drives the service's cached-sandbox store over the real
+  `lns artifact` drives the service's cache over the real
   Unix-socket IPC. Pulling needs registry network access, so this
   feature pins the host-side wiring of the offline verbs through real
   binaries: ls renders the cache table, rm refuses unknown sandboxes, and
@@ -10,7 +10,7 @@ Feature: cached sandbox management end to end
     And the Lens Sandbox service is running in that home
 
   Scenario: listing an empty cache renders only the table header
-    When I run "lns sandbox ls"
+    When I run "lns artifact ls"
     Then the exit code is 0
     And the output contains "ARTIFACT"
     And the output contains "KIND"
@@ -19,16 +19,16 @@ Feature: cached sandbox management end to end
     And the output contains "HOLDER"
 
   Scenario: removing a sandbox that is not cached fails cleanly
-    When I run "lns sandbox rm registry.example.test/absent:1"
+    When I run "lns artifact rm registry.example.test/absent:1"
     Then the exit code is non-zero
     And the output contains "no such image"
 
   Scenario: pruning an empty cache reclaims nothing
-    When I run "lns sandbox prune --force"
+    When I run "lns artifact prune --force"
     Then the exit code is 0
     And the output contains "reclaimed 0 B"
 
   Scenario: with no terminal to ask at, prune refuses rather than assuming
-    When I run "lns sandbox prune"
+    When I run "lns artifact prune"
     Then the exit code is non-zero
     And the output contains "pass --force to confirm"
