+++ crates/lns-cli/src/output.rs
@@ -123,6 +123,22 @@ fn write_row(out: &mut dyn Write, cells: &[String], widths: &[usize]) -> std::io
     writeln!(out)
 }
 
+/// The size column every table renders through, in the binary units the JSON carries raw.
+pub fn format_bytes(n: u64) -> String {
+    const KIB: u64 = 1024;
+    const MIB: u64 = KIB * 1024;
+    const GIB: u64 = MIB * 1024;
+    if n >= GIB {
+        format!("{:.1} GiB", n as f64 / GIB as f64)
+    } else if n >= MIB {
+        format!("{:.1} MiB", n as f64 / MIB as f64)
+    } else if n >= KIB {
+        format!("{:.1} KiB", n as f64 / KIB as f64)
+    } else {
+        format!("{n} B")
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
