+++ crates/lns-cli/src/sandbox/mod.rs
@@ -1,23 +1,15 @@
-use std::path::PathBuf;
-
 use anyhow::{Context, Result, bail};
 use lns_ipc::{
     Request, Response, RunDetails, RunStatsInfo, WireFrame, decode_wire_frame_from_bytes,
     read_frame_bytes_async,
 };
-use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
+use tokio::io::{AsyncRead, AsyncWriteExt};
 
 use crate::cli::{DetachChord, ExecArgs, KillArgs, RunArgs, parse_detach_keys_arg};
 use crate::command::{CommandSpec, subcommand};
-use crate::service::client::BoxFuture;
+pub use crate::service::client::{BoxFuture, SandboxService, TermInfo};
 
-pub mod author;
-pub mod distribute;
-pub mod fileset;
-pub mod mixin_plan;
 pub mod real;
-#[cfg(test)]
-pub(crate) mod test_support;
 
 #[derive(clap::Args)]
 pub struct SandboxArgs {
@@ -25,195 +17,36 @@ pub struct SandboxArgs {
     pub command: SandboxCommand,
 }
 
+/// Only the verbs that act on a sandbox — the document one is made from is `lns artifact`'s.
 #[derive(clap::Subcommand)]
 pub enum SandboxCommand {
-    #[command(about = "Scaffold a document in this directory; `--kind` chooses which.")]
-    Init(SandboxInitArgs),
-    #[command(about = "Validate ./lns.yaml — schema, cross-field, and secret checks, offline.")]
-    Validate(SandboxValidateArgs),
-    #[command(
-        about = "Build ./lns.yaml and upload it to a registry as a sandbox artifact, in one step."
-    )]
-    Push(SandboxPushArgs),
-    #[command(about = "Fetch a published sandbox and its base image into the local cache.")]
-    Pull(SandboxPullArgs),
-    #[command(about = "Add a tag to a cached sandbox within its current repository.")]
-    Tag(SandboxTagArgs),
-    #[command(about = "List running sandboxes with their CPU and memory (`docker ps`-style).")]
-    Ps(PsArgs),
-    #[command(
-        visible_alias = "list",
-        about = "List cached sandboxes (pulled or built) in the local store."
-    )]
-    Ls(LsArgs),
     #[command(about = "Run a sandbox in a microVM (the top-level `lns run`).")]
     Run(Box<RunArgs>),
-    #[command(about = "Run another command against a running run.")]
+    #[command(about = "Run another command inside a running sandbox.")]
     Exec(ExecArgs),
-    #[command(about = "Send a signal to a running run (`docker kill`-style).")]
+    #[command(about = "Send a signal to a running sandbox (`docker kill`-style).")]
     Kill(KillArgs),
-    #[command(about = "Stop a run gracefully: SIGTERM, then SIGKILL once the timeout passes.")]
+    #[command(about = "Stop a sandbox gracefully: SIGTERM, then SIGKILL once the timeout passes.")]
     Stop(SandboxStopArgs),
-    #[command(about = "Print a run's captured output; `-f` streams until the run exits.")]
+    #[command(about = "Print a sandbox's captured output; `-f` streams until the workload exits.")]
     Logs(SandboxLogsArgs),
-    #[command(about = "Re-attach to a running run's output (detach chord to leave again).")]
+    #[command(about = "Re-attach to a running sandbox's output (detach chord to leave again).")]
     Attach(SandboxAttachArgs),
     #[command(
-        about = "Inspect a sandbox: a running run's live state as JSON, a cached artifact's kind and definition, or a local lns.yaml's effective definition (offline)."
+        visible_alias = "list",
+        about = "List running sandboxes with their state, CPU and memory."
     )]
+    Ls(SandboxLsArgs),
+    #[command(about = "Print one sandbox's live state and launch configuration.")]
     Inspect(SandboxInspectArgs),
-    #[command(
-        about = "Remove a cached sandbox and free its now-unreferenced layers; refuses a running one."
-    )]
+    #[command(about = "Remove a sandbox: its record and its writable layer.")]
     Rm(SandboxRmArgs),
-    #[command(
-        about = "Remove every cached sandbox not held by a running one and reclaim the provisioned tool cache when no sandbox is live."
-    )]
-    Prune(SandboxPruneArgs),
 }
 
 #[derive(clap::Args)]
-pub struct PsArgs {
+pub struct SandboxLsArgs {
     #[command(flatten)]
     pub output: crate::output::OutputArgs,
-}
-
-#[derive(clap::Args)]
-pub struct LsArgs {
-    #[arg(
-        long,
-        value_enum,
-        value_name = "KIND",
-        help = "List only the cached entries of this kind."
-    )]
-    pub kind: Option<CachedKindFilter>,
-
-    #[command(flatten)]
-    pub output: crate::output::OutputArgs,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, clap::ValueEnum)]
-#[value(rename_all = "lower")]
-pub enum CachedKindFilter {
-    Image,
-    Sandbox,
-}
-
-impl CachedKindFilter {
-    fn matches(self, kind: lns_ipc::CachedKind) -> bool {
-        matches!(
-            (self, kind),
-            (CachedKindFilter::Image, lns_ipc::CachedKind::Image)
-                | (CachedKindFilter::Sandbox, lns_ipc::CachedKind::Sandbox)
-        )
-    }
-}
-
-#[derive(clap::Args)]
-pub struct SandboxInitArgs {
-    #[arg(
-        long,
-        value_enum,
-        default_value_t = DocumentKind::Sandbox,
-        value_name = "KIND",
-        help = "Which document to scaffold."
-    )]
-    pub kind: DocumentKind,
-
-    #[arg(
-        short = 'f',
-        long = "file",
-        value_name = "FILE",
-        help = "File to write instead of ./lns.yaml. Refuses to overwrite either way."
-    )]
-    pub file: Option<PathBuf>,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq, clap::ValueEnum)]
-#[value(rename_all = "lower")]
-pub enum DocumentKind {
-    Sandbox,
-    Mixin,
-}
-
-impl DocumentKind {
-    pub fn as_artifact_kind(self) -> lns_artifact::spec::Kind {
-        match self {
-            DocumentKind::Sandbox => lns_artifact::spec::Kind::Sandbox,
-            DocumentKind::Mixin => lns_artifact::spec::Kind::Mixin,
-        }
-    }
-}
-
-#[derive(clap::Args)]
-pub struct SandboxValidateArgs {
-    #[arg(
-        long,
-        value_enum,
-        value_name = "KIND",
-        help = "Also require the document to be this kind."
-    )]
-    pub kind: Option<DocumentKind>,
-
-    #[arg(
-        short = 'f',
-        long = "file",
-        value_name = "FILE",
-        help = "Definition file to validate instead of ./lns.yaml; its directory roots the definition's relative filesets."
-    )]
-    pub file: Option<PathBuf>,
-}
-
-#[derive(clap::Args)]
-pub struct SandboxPushArgs {
-    #[arg(
-        value_name = "REF",
-        help = "Registry reference to publish the sandbox at, e.g. ghcr.io/team/hermes:1.4.0; a bare reference resolves against the `run.registry` default, else hub.lns.run."
-    )]
-    pub reference: String,
-    #[arg(
-        long = "dry-run",
-        help = "Validate, pack, and build everything push would upload, print the digests, and upload nothing."
-    )]
-    pub dry_run: bool,
-    #[arg(
-        long = "yes",
-        help = "Publish the local mixins this document names without prompting."
-    )]
-    pub assume_yes: bool,
-    #[arg(
-        short = 'f',
-        long = "file",
-        value_name = "FILE",
-        help = "Definition file to publish instead of ./lns.yaml; its directory roots the definition's relative filesets."
-    )]
-    pub file: Option<PathBuf>,
-}
-
-#[derive(clap::Args)]
-pub struct SandboxPullArgs {
-    #[arg(
-        value_name = "REF",
-        help = "Published sandbox reference to fetch, e.g. ghcr.io/team/hermes:1.4.0; a bare reference resolves against the `run.registry` default, else hub.lns.run."
-    )]
-    pub reference: String,
-    #[arg(
-        long = "yes",
-        help = "Accept publisher-declared tool installation without prompting."
-    )]
-    pub assume_yes: bool,
-}
-
-#[derive(clap::Args)]
-pub struct SandboxTagArgs {
-    #[arg(value_name = "SOURCE", help = "Cached sandbox to re-reference.")]
-    pub from: String,
-
-    #[arg(
-        value_name = "TARGET",
-        help = "New tag in the source sandbox's registry and repository."
-    )]
-    pub to: String,
 }
 
 #[derive(clap::Args)]
@@ -270,26 +103,10 @@ pub struct SandboxAttachArgs {
 #[derive(clap::Args)]
 pub struct SandboxInspectArgs {
     #[arg(
-        value_name = "TARGET",
-        help = "A running run's id/name (live state), a cached sandbox reference (its definition), or a path to a local definition (., lns.yaml, ./dir, ./lns.dev.yaml — rendered offline). Omit to render ./lns.yaml."
+        value_name = "RUN",
+        help = "Sandbox id or name surfaced by `lns sandbox ls`."
     )]
-    pub run: Option<String>,
-
-    #[arg(
-        short = 'f',
-        long = "file",
-        value_name = "FILE",
-        conflicts_with = "run",
-        help = "Definition file to render instead of ./lns.yaml, offline. Cannot be combined with TARGET."
-    )]
-    pub file: Option<PathBuf>,
-
-    #[arg(
-        long = "mixin",
-        value_name = "REF",
-        help = "Resolve this mixin into the sandbox before rendering it, as `lns run --mixin` would. Repeatable."
-    )]
-    pub mixins: Vec<String>,
+    pub run: String,
 
     #[command(flatten)]
     pub output: crate::output::OutputArgs,
@@ -304,31 +121,10 @@ pub struct SandboxRmArgs {
     pub run: String,
 }
 
-#[derive(clap::Args)]
-pub struct SandboxPruneArgs {
-    #[arg(
-        short = 'f',
-        long,
-        default_value_t = false,
-        help = "Remove unused cached sandboxes, and the provisioned tool cache when none is live, without asking."
-    )]
-    pub force: bool,
-}
-
-/// Qualifies the registry coordinate a distribution verb addresses with `registry`, or the built-in default when nothing is configured; the local-cache verbs take an id-or-ref the service resolves, so they are left alone.
-pub fn apply_registry_default(command: &mut SandboxCommand, registry: Option<&str>) {
-    let reference = match command {
-        SandboxCommand::Push(args) => &mut args.reference,
-        SandboxCommand::Pull(args) => &mut args.reference,
-        _ => return,
-    };
-    *reference = crate::config::resolve_default_registry(reference, registry);
-}
-
 pub fn augment(app: clap::Command) -> clap::Command {
     app.subcommand(
         subcommand::<SandboxArgs>("sandbox")
-            .about("The sandbox — author, distribute, run, and manage it (the complete surface).")
+            .about("The sandbox a document becomes when it runs — create, watch, and remove it.")
             .mut_subcommand("run", crate::run::long_help_only),
     )
 }
@@ -349,25 +145,10 @@ pub const SPEC: CommandSpec = CommandSpec {
     owns_terminal: verb_owns_terminal,
 };
 
-pub fn augment_init(app: clap::Command) -> clap::Command {
-    app.subcommand(
-        subcommand::<SandboxInitArgs>("init")
-            .about("Scaffold a document in this directory (shortcut for `lns sandbox init`)."),
-    )
-}
-
-pub const INIT_SPEC: CommandSpec = CommandSpec {
-    name: "init",
-    augment: augment_init,
-    run: real::run_init,
-    announces_update_check: true,
-    owns_terminal: crate::command::never_owns_terminal,
-};
-
 pub fn augment_ps(app: clap::Command) -> clap::Command {
     app.subcommand(
-        subcommand::<PsArgs>("ps").about(
-            "List running sandboxes with their CPU and memory (shortcut for `lns sandbox ps`).",
+        subcommand::<SandboxLsArgs>("ps").about(
+            "List running sandboxes with their CPU and memory (shortcut for `lns sandbox ls`).",
         ),
     )
 }
@@ -423,22 +204,6 @@ shortcut_spec!(
     "Send a signal to a running sandbox (shortcut for `lns sandbox kill`)."
 );
 shortcut_spec!(
-    augment_rm,
-    RM_SPEC,
-    SandboxRmArgs,
-    "rm",
-    real::run_rm,
-    "Remove a sandbox (shortcut for `lns sandbox rm`)."
-);
-shortcut_spec!(
-    augment_inspect,
-    INSPECT_SPEC,
-    SandboxInspectArgs,
-    "inspect",
-    real::run_inspect,
-    "Inspect a sandbox (shortcut for `lns sandbox inspect`)."
-);
-shortcut_spec!(
     augment_logs,
     LOGS_SPEC,
     SandboxLogsArgs,
@@ -456,86 +221,31 @@ shortcut_spec!(
     "Re-attach to a running sandbox (shortcut for `lns sandbox attach`).",
     crate::command::always_owns_terminal
 );
-shortcut_spec!(
-    augment_push,
-    PUSH_SPEC,
-    SandboxPushArgs,
-    "push",
-    real::run_push,
-    "Build and publish ./lns.yaml as a sandbox artifact (shortcut for `lns sandbox push`)."
-);
-shortcut_spec!(
-    augment_pull,
-    PULL_SPEC,
-    SandboxPullArgs,
-    "pull",
-    real::run_pull,
-    "Fetch a published sandbox into the cache (shortcut for `lns sandbox pull`)."
-);
-shortcut_spec!(
-    augment_tag,
-    TAG_SPEC,
-    SandboxTagArgs,
-    "tag",
-    real::run_tag,
-    "Re-reference a cached sandbox (shortcut for `lns sandbox tag`)."
-);
 
-pub trait SandboxService: Send + Sync {
-    type Stream: AsyncRead + AsyncWrite + Unpin + Send + 'static;
-    fn one_shot(&self, request: Request) -> BoxFuture<'_, Result<Response>>;
-    fn open_stream(&self, request: Request) -> BoxFuture<'_, Result<Self::Stream>>;
-    fn aux_socket(&self) -> Option<PathBuf>;
-    fn load_policy(&self, path: &str) -> Option<serde_json::Value>;
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub struct TermInfo {
-    pub stdin_is_tty: bool,
-    pub stdout_is_terminal: bool,
-}
-
-pub async fn run_with_writers<S, I, W, O, E>(
+pub async fn run_with_writers<S, W, O, E>(
     cmd: &SandboxCommand,
     svc: &S,
     term: TermInfo,
-    input: &mut I,
     out: &mut W,
     stdout: &mut O,
     stderr: &mut E,
 ) -> Result<i32>
 where
     S: SandboxService,
-    I: std::io::BufRead,
     W: std::io::Write,
     O: AsyncWriteExt + Unpin,
     E: AsyncWriteExt + Unpin,
 {
     match cmd {
-        SandboxCommand::Init(_) | SandboxCommand::Validate(_) => {
-            bail!("author commands run offline, not through the service dispatch")
-        }
-        SandboxCommand::Push(_) => {
-            bail!("push builds and uploads locally, not through the service dispatch")
-        }
-        SandboxCommand::Pull(args) => pull(svc, args, term, input, out, stderr).await,
-        SandboxCommand::Tag(args) => tag(svc, args, out).await,
-        SandboxCommand::Ps(args) => ps(svc, args, out).await,
-        SandboxCommand::Ls(args) => ls(svc, args, out).await,
+        SandboxCommand::Ls(args) => ps(svc, args, out).await,
         SandboxCommand::Kill(args) => kill(svc, args, out).await,
         SandboxCommand::Run(_) => bail!("sandbox run is dispatched on its own interactive path"),
         SandboxCommand::Exec(_) => bail!("sandbox exec is dispatched on its own interactive path"),
         SandboxCommand::Stop(args) => stop(svc, args, out).await,
-        SandboxCommand::Inspect(args) => {
-            let Some(target) = &args.run else {
-                bail!("a local definition inspect runs offline, not through the service dispatch")
-            };
-            inspect(svc, target, &args.mixins, args.output.format, out).await
-        }
+        SandboxCommand::Inspect(args) => inspect(svc, &args.run, args.output.format, out).await,
         SandboxCommand::Logs(args) => logs(svc, args, stdout, stderr).await,
         SandboxCommand::Attach(args) => attach(svc, args, term, stdout, stderr).await,
         SandboxCommand::Rm(args) => rm(svc, args, out).await,
-        SandboxCommand::Prune(args) => prune(svc, args, term, input, out).await,
     }
 }
 
@@ -543,124 +253,9 @@ pub(crate) fn run_label(run: &str) -> String {
     run.to_string()
 }
 
-async fn pull<I, W, E>(
-    svc: &impl SandboxService,
-    args: &SandboxPullArgs,
-    term: TermInfo,
-    input: &mut I,
-    out: &mut W,
-    stderr: &mut E,
-) -> Result<i32>
-where
-    I: std::io::BufRead,
-    W: std::io::Write,
-    E: AsyncWriteExt + Unpin,
-{
-    let inspection = svc
-        .one_shot(Request::InspectImage {
-            image: args.reference.clone(),
-            mixins: Vec::new(),
-            decisions: None,
-        })
-        .await?;
-    // A mixin pull installs nothing: it caches documents, so there is no effect to consent to and its tools are disclosed where they are installed.
-    let (digest, tools) = match inspection {
-        Response::ImageInspected {
-            inspection: lns_ipc::ArtifactInspection::Sandbox(view),
-        } if !view.digest.is_empty() => (view.digest, view.tools),
-        Response::ImageInspected {
-            inspection: lns_ipc::ArtifactInspection::Mixin(view),
-        } if !view.digest.is_empty() => (view.digest, Vec::new()),
-        Response::ImageInspected {
-            inspection:
-                lns_ipc::ArtifactInspection::Mixin(_) | lns_ipc::ArtifactInspection::Sandbox(_),
-        } => bail!(
-            "the registry did not provide a digest for {}",
-            args.reference
-        ),
-        Response::ImageInspected {
-            inspection: lns_ipc::ArtifactInspection::Image(_),
-        } => bail!(
-            "{} is an OCI image, not a published sandbox",
-            args.reference
-        ),
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    };
-    crate::run::pull_confirm::confirm_pulled_effects(
-        &crate::run::pull_confirm::PulledEffects {
-            reference: &args.reference,
-            binds: &[],
-            volumes: &[],
-            filesets: &[],
-            tools: &tools,
-            // A pull runs no script: scripts run at boot, and `lns run` discloses them there.
-            scripts: &[],
-        },
-        args.assume_yes,
-        term.stdin_is_tty,
-        input,
-        out,
-    )?;
-    let response = svc
-        .one_shot(Request::PullImage {
-            image: args.reference.clone(),
-            expected_digest: digest,
-        })
-        .await?;
-    match response {
-        Response::MixinPulled {
-            reference,
-            digest,
-            cached_mixins,
-        } => {
-            writeln!(out, "pulled {reference}")?;
-            writeln!(out, "digest: {digest}")?;
-            if cached_mixins > 0 {
-                writeln!(out, "cached {cached_mixins} mixin(s) it layers on")?;
-            }
-            Ok(0)
-        }
-        Response::ImagePulled { image, warnings } => {
-            writeln!(out, "pulled {}", image.reference)?;
-            writeln!(out, "digest: {}", image.digest)?;
-            for warning in warnings {
-                stderr
-                    .write_all(format!("warning: {warning}\n").as_bytes())
-                    .await?;
-                stderr.flush().await?;
-            }
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
-async fn tag<W: std::io::Write>(
-    svc: &impl SandboxService,
-    args: &SandboxTagArgs,
-    out: &mut W,
-) -> Result<i32> {
-    let response = svc
-        .one_shot(Request::TagImage {
-            from: args.from.clone(),
-            to: args.to.clone(),
-        })
-        .await?;
-    match response {
-        Response::ImageTagged { from, to } => {
-            writeln!(out, "tagged {from} as {to}")?;
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
 async fn ps<W: std::io::Write>(
     svc: &impl SandboxService,
-    args: &PsArgs,
+    args: &SandboxLsArgs,
     out: &mut W,
 ) -> Result<i32> {
     let running = match svc.one_shot(Request::ListRuns).await? {
@@ -723,7 +318,11 @@ impl PsRow {
     fn memory_cell(&self) -> String {
         match (self.mem_used_bytes, self.mem_total_bytes) {
             (Some(used), Some(total)) => {
-                format!("{} / {}", format_bytes(used), format_bytes(total))
+                format!(
+                    "{} / {}",
+                    crate::output::format_bytes(used),
+                    crate::output::format_bytes(total)
+                )
             }
             _ => NO_SAMPLE.to_string(),
         }
@@ -741,79 +340,6 @@ impl crate::output::TableRow for PsRow {
             self.cpu_permille
                 .map_or_else(|| NO_SAMPLE.to_string(), format_permille),
             self.memory_cell(),
-        ]
-    }
-}
-
-async fn ls<W: std::io::Write>(
-    svc: &impl SandboxService,
-    args: &LsArgs,
-    out: &mut W,
-) -> Result<i32> {
-    match svc.one_shot(Request::ListImages).await? {
-        Response::ImageList { mut images } => {
-            images.sort_by(|a, b| a.reference.cmp(&b.reference));
-            let rows: Vec<SandboxRow> = images
-                .iter()
-                .filter(|image| args.kind.is_none_or(|kind| kind.matches(image.kind)))
-                .map(SandboxRow::new)
-                .collect();
-            crate::output::emit(args.output.format, &rows, out)?;
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
-#[derive(serde::Serialize)]
-#[serde(rename_all = "camelCase")]
-struct SandboxRow {
-    reference: String,
-    kind: String,
-    digest: String,
-    size_bytes: u64,
-    layers: u32,
-    pulled: String,
-    in_use_by: Option<String>,
-}
-
-impl SandboxRow {
-    fn new(image: &lns_ipc::ImageInfo) -> Self {
-        Self {
-            reference: image.reference.clone(),
-            kind: image.kind.as_str().to_string(),
-            digest: image.digest.clone(),
-            size_bytes: image.size_bytes,
-            layers: image.layers,
-            pulled: image.pulled.clone(),
-            in_use_by: image.in_use_by.clone(),
-        }
-    }
-}
-
-/// The digest column is the short form the tables elsewhere use, since the whole one crowds out every other column and the JSON carries it in full.
-fn short_digest(digest: &str) -> String {
-    let hex = digest.strip_prefix("sha256:").unwrap_or(digest);
-    hex.char_indices()
-        .nth(12)
-        .map_or(hex, |(i, _)| &hex[..i])
-        .to_string()
-}
-
-impl crate::output::TableRow for SandboxRow {
-    const HEADERS: &'static [&'static str] = &["ARTIFACT", "KIND", "DIGEST", "SIZE", "HOLDER"];
-
-    fn cells(&self) -> Vec<String> {
-        vec![
-            self.reference.clone(),
-            self.kind.clone(),
-            short_digest(&self.digest),
-            format_bytes(self.size_bytes),
-            match &self.in_use_by {
-                Some(run) => format!("run {}", lns_ipc::short_run_id(run)),
-                None => "-".to_string(),
-            },
         ]
     }
 }
@@ -895,12 +421,8 @@ async fn rm<W: std::io::Write>(
                 args.run
             )
         }
-        // An exited run is a spent record.
+        // An exited sandbox is a spent record.
         Response::RunInspect { .. } => remove_run(svc, &args.run, out).await,
-        // A reference the service knows no run for is a cached sandbox; any other error is a real failure, not a miss.
-        Response::Error { message } if is_unknown_run(&message) => {
-            remove_cached(svc, &args.run, out).await
-        }
         Response::Error { message } => bail!("daemon error: {message}"),
         other => bail!("unexpected response from daemon: {other:?}"),
     }
@@ -926,101 +448,9 @@ async fn remove_run<W: std::io::Write>(
     }
 }
 
-async fn remove_cached<W: std::io::Write>(
-    svc: &impl SandboxService,
-    reference: &str,
-    out: &mut W,
-) -> Result<i32> {
-    match svc
-        .one_shot(Request::RemoveImage {
-            image: reference.to_string(),
-        })
-        .await?
-    {
-        Response::ImageRemoved {
-            reference,
-            reclaimed_bytes,
-        } => {
-            writeln!(out, "removed {reference}")?;
-            writeln!(
-                out,
-                "freed {} of base-image layers",
-                format_bytes(reclaimed_bytes)
-            )?;
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
-async fn prune<I: std::io::BufRead, W: std::io::Write>(
-    svc: &impl SandboxService,
-    args: &SandboxPruneArgs,
-    term: TermInfo,
-    input: &mut I,
-    out: &mut W,
-) -> Result<i32> {
-    if !args.force {
-        if !term.stdin_is_tty {
-            bail!(
-                "this removes every cached sandbox not held by a running one and, when none is live, the provisioned tool cache; there is no terminal to ask at, so pass --force to confirm"
-            );
-        }
-        if !confirm_prune(input, out)? {
-            return Ok(0);
-        }
-    }
-    match svc.one_shot(Request::PruneImages).await? {
-        Response::ImagesPruned {
-            mut removed,
-            reclaimed_bytes,
-        } => {
-            removed.sort_unstable();
-            for reference in &removed {
-                writeln!(out, "removed {reference}")?;
-            }
-            writeln!(out, "reclaimed {}", format_bytes(reclaimed_bytes))?;
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
-fn confirm_prune<I: std::io::BufRead, W: std::io::Write>(
-    input: &mut I,
-    out: &mut W,
-) -> Result<bool> {
-    write!(
-        out,
-        "This removes every cached sandbox not held by a running one and, when none is live, the provisioned tool cache. Continue? [y/N] "
-    )?;
-    out.flush()?;
-    let mut line = String::new();
-    input.read_line(&mut line)?;
-    let answer = line.trim();
-    let yes = answer.eq_ignore_ascii_case("y") || answer.eq_ignore_ascii_case("yes");
-    if !yes {
-        writeln!(out, "Aborted.")?;
-    }
-    Ok(yes)
-}
-
-/// `--mixin` composes a document before it boots, and neither a run that has already booted nor a file rendered offline can honour it.
-pub fn refuse_mixins_unless_published(mixins: &[String]) -> Result<()> {
-    if mixins.is_empty() {
-        return Ok(());
-    }
-    bail!(
-        "--mixin applies to a sandbox reference: a live run has already booted with what it merged, and a local definition renders offline, without resolving anything"
-    )
-}
-
 async fn inspect<W: std::io::Write>(
     svc: &impl SandboxService,
     target: &str,
-    mixins: &[String],
     format: crate::output::Format,
     out: &mut W,
 ) -> Result<i32> {
@@ -1031,7 +461,6 @@ async fn inspect<W: std::io::Write>(
         .await?
     {
         Response::RunInspect { details } => {
-            refuse_mixins_unless_published(mixins)?;
             let policy = details
                 .config
                 .policy_path
@@ -1040,203 +469,9 @@ async fn inspect<W: std::io::Write>(
             render_inspect(&details, policy, format, out)?;
             Ok(0)
         }
-        // A reference the service knows no run for is a cached artifact; any other error is a real failure, not a miss.
-        Response::Error { message } if is_unknown_run(&message) => {
-            refuse_format_for_a_document(format)?;
-            inspect_cached(svc, target, mixins, out).await
-        }
         Response::Error { message } => bail!("daemon error: {message}"),
         other => bail!("unexpected response from daemon: {other:?}"),
     }
-}
-
-async fn inspect_cached<W: std::io::Write>(
-    svc: &impl SandboxService,
-    reference: &str,
-    mixins: &[String],
-    out: &mut W,
-) -> Result<i32> {
-    match svc
-        .one_shot(Request::InspectImage {
-            image: reference.to_string(),
-            mixins: mixins.to_vec(),
-            decisions: None,
-        })
-        .await?
-    {
-        Response::ImageInspected { inspection } => {
-            render_cached_inspect(&inspection, mixins, out)?;
-            Ok(0)
-        }
-        Response::Error { message } => bail!("daemon error: {message}"),
-        other => bail!("unexpected response from daemon: {other:?}"),
-    }
-}
-
-fn render_cached_inspect<W: std::io::Write>(
-    inspection: &lns_ipc::ArtifactInspection,
-    typed: &[String],
-    out: &mut W,
-) -> Result<()> {
-    match inspection {
-        lns_ipc::ArtifactInspection::Sandbox(view) => {
-            writeln!(out, "kind: sandbox")?;
-            writeln!(out, "reference: {}", view.reference)?;
-            if !view.digest.is_empty() {
-                writeln!(out, "digest: {}", view.digest)?;
-            }
-            writeln!(out, "image: {}", view.image)?;
-            for mixin in
-                crate::run::summary::mixin_display(&view.mixins, typed, &view.pinned_mixins)
-            {
-                writeln!(out, "mixin: {mixin}")?;
-            }
-            if let Some(user) = &view.user {
-                writeln!(out, "user: {user}")?;
-            }
-            if let Some(workdir) = &view.workdir {
-                writeln!(out, "workdir: {workdir}")?;
-            }
-            for entry in &view.env {
-                writeln!(out, "env: {entry}")?;
-            }
-            render_mounts(out, &view.mounts)?;
-            if !view.ports.is_empty() {
-                writeln!(out, "ports: {}", declared_ports_line(&view.ports))?;
-            }
-            render_filesets(out, &view.filesets)?;
-            render_connectors(out, &view.connectors)?;
-            for credential in &view.credentials {
-                writeln!(out, "credential: {}", credential_disclosure(credential))?;
-            }
-            for tool in &view.tools {
-                writeln!(out, "tool: {tool}")?;
-            }
-            render_scripts(out, &view.scripts)?;
-            render_policy_flags(out, &view.policy_flags)?;
-        }
-        lns_ipc::ArtifactInspection::Mixin(view) => {
-            writeln!(out, "kind: mixin")?;
-            writeln!(out, "reference: {}", view.reference)?;
-            if !view.digest.is_empty() {
-                writeln!(out, "digest: {}", view.digest)?;
-            }
-            for mixin in &view.mixins {
-                writeln!(out, "mixin: {mixin}")?;
-            }
-            for entry in &view.env {
-                writeln!(out, "env: {entry}")?;
-            }
-            render_mounts(out, &view.mounts)?;
-            if !view.ports.is_empty() {
-                writeln!(out, "ports: {}", declared_ports_line(&view.ports))?;
-            }
-            render_filesets(out, &view.filesets)?;
-            for credential in &view.credentials {
-                writeln!(out, "credential: {}", credential_disclosure(credential))?;
-            }
-            for tool in &view.tools {
-                writeln!(out, "tool: {tool}")?;
-            }
-            render_scripts(out, &view.scripts)?;
-            render_policy_flags(out, &view.policy_flags)?;
-        }
-        lns_ipc::ArtifactInspection::Image(view) => {
-            writeln!(out, "kind: image")?;
-            writeln!(out, "reference: {}", view.reference)?;
-            writeln!(out, "digest: {}", view.digest)?;
-        }
-    }
-    Ok(())
-}
-
-/// One renderer for both kinds, since a mixin's mounts read exactly as a sandbox's do.
-fn render_mounts<W: std::io::Write>(out: &mut W, mounts: &[lns_ipc::SandboxMount]) -> Result<()> {
-    for mount in mounts {
-        let kind = match mount.kind {
-            lns_ipc::SandboxMountKind::Bind => "bind",
-            lns_ipc::SandboxMountKind::Volume => "volume",
-        };
-        let mode = if mount.read_only { " (read-only)" } else { "" };
-        writeln!(
-            out,
-            "mount: {kind} {} -> {}{mode}",
-            mount.source, mount.target
-        )?;
-    }
-    Ok(())
-}
-
-fn render_filesets<W: std::io::Write>(
-    out: &mut W,
-    filesets: &[lns_ipc::SandboxFileset],
-) -> Result<()> {
-    for fileset in filesets {
-        let source = crate::run::summary::fileset_view_source_display(fileset);
-        let owner = crate::run::summary::fileset_view_owner_display(fileset.owner);
-        writeln!(
-            out,
-            "fileset: {source} -> {} (owner: {owner})",
-            fileset.guest_path
-        )?;
-    }
-    Ok(())
-}
-
-fn declared_ports_line(ports: &[lns_ipc::SandboxPort]) -> String {
-    ports
-        .iter()
-        .map(|port| match port.host {
-            Some(host) => format!("{host}:{}", port.container),
-            None => port.container.to_string(),
-        })
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-/// What a declared credential discloses: the variable the workload reads, and every destination its value may travel to — the two things a reader has to weigh before running it.
-pub(crate) fn credential_disclosure(credential: &lns_spec::Credential) -> String {
-    if credential.injections.is_empty() {
-        return format!("{} (travels nowhere)", credential.env_var);
-    }
-    let domains: Vec<&str> = credential
-        .injections
-        .iter()
-        .map(|injection| injection.domain.as_str())
-        .collect();
-    format!("{} -> {}", credential.env_var, domains.join(", "))
-}
-
-fn render_connectors<W: std::io::Write>(out: &mut W, connectors: &[String]) -> Result<()> {
-    for id in connectors {
-        writeln!(out, "connector: {id}")?;
-    }
-    Ok(())
-}
-
-/// Inspect is where a script's body is printed whole; the launch summary elides it so a long one cannot bury the rest of the approval.
-fn render_scripts<W: std::io::Write>(
-    out: &mut W,
-    scripts: &[lns_ipc::SandboxScript],
-) -> Result<()> {
-    for script in scripts {
-        let user = script.user.as_deref().unwrap_or("the workload user");
-        writeln!(out, "script: {} (runs as {user})", script.when)?;
-        if let Some(description) = &script.description {
-            writeln!(out, "  {description}")?;
-        }
-        for line in script.run.lines() {
-            writeln!(out, "  | {line}")?;
-        }
-    }
-    Ok(())
-}
-
-fn render_policy_flags<W: std::io::Write>(out: &mut W, flags: &[String]) -> Result<()> {
-    for flag in flags {
-        writeln!(out, "⚠ {flag}")?;
-    }
-    Ok(())
 }
 
 fn policy_doc(path: &str, loaded: Option<serde_json::Value>) -> serde_json::Value {
@@ -1244,16 +479,6 @@ fn policy_doc(path: &str, loaded: Option<serde_json::Value>) -> serde_json::Valu
         Some(doc) => serde_json::json!({ "path": path, "policy": doc }),
         None => serde_json::json!({ "path": path, "error": "policy file could not be read" }),
     }
-}
-
-/// `--format` answers for a sandbox's live state (§3.2). A document renders as the prose its author wrote, so asking for a shape it has no answer for is a refusal rather than a flag silently doing nothing.
-fn refuse_format_for_a_document(format: crate::output::Format) -> Result<()> {
-    if format == crate::output::Format::Json {
-        bail!(
-            "--format answers for a running sandbox; a cached artifact and a local document render as themselves"
-        );
-    }
-    Ok(())
 }
 
 fn render_inspect<W: std::io::Write>(
@@ -1326,7 +551,7 @@ fn inspect_fields(details: &RunDetails) -> Vec<(&'static str, String)> {
         ("CPUS", details.config.cpus.to_string()),
         (
             "MEM",
-            format_bytes(details.config.mem_mib as u64 * 1024 * 1024),
+            crate::output::format_bytes(details.config.mem_mib as u64 * 1024 * 1024),
         ),
         (
             "WORKDIR",
@@ -1345,21 +570,6 @@ fn inspect_fields(details: &RunDetails) -> Vec<(&'static str, String)> {
 
 fn format_permille(permille: u32) -> String {
     format!("{}.{}%", permille / 10, permille % 10)
-}
-
-fn format_bytes(n: u64) -> String {
-    const KIB: u64 = 1024;
-    const MIB: u64 = KIB * 1024;
-    const GIB: u64 = MIB * 1024;
-    if n >= GIB {
-        format!("{:.1} GiB", n as f64 / GIB as f64)
-    } else if n >= MIB {
-        format!("{:.1} MiB", n as f64 / MIB as f64)
-    } else if n >= KIB {
-        format!("{:.1} KiB", n as f64 / KIB as f64)
-    } else {
-        format!("{n} B")
-    }
 }
 
 async fn logs<S, O, E>(
@@ -1458,153 +668,41 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_service::{CannedService, stream_with};
     use lns_ipc::encode_frame;
-    use std::sync::{Arc, Mutex};
 
-    struct CannedService {
-        response: Response,
-        stats_response: Option<Response>,
-        inspect_image_response: Option<Response>,
-        remove_image_response: Option<Response>,
-        remove_run_response: Option<Response>,
-        frames: Vec<Vec<u8>>,
-        requests: Arc<Mutex<Vec<Request>>>,
-    }
-
-    impl CannedService {
-        fn new(response: Response) -> Self {
-            Self {
-                response,
-                stats_response: None,
-                inspect_image_response: None,
-                remove_image_response: None,
-                remove_run_response: None,
-                frames: Vec::new(),
-                requests: Arc::new(Mutex::new(Vec::new())),
-            }
-        }
-
-        fn with_stats(response: Response, stats_response: Response) -> Self {
-            Self {
-                stats_response: Some(stats_response),
-                ..Self::new(response)
-            }
-        }
-
-        fn with_inspect_image(run_response: Response, image_response: Response) -> Self {
-            Self {
-                inspect_image_response: Some(image_response),
-                ..Self::new(run_response)
-            }
-        }
-
-        fn with_remove_image(run_response: Response, remove_response: Response) -> Self {
-            Self {
-                remove_image_response: Some(remove_response),
-                ..Self::new(run_response)
-            }
-        }
-
-        fn with_remove_run(run_response: Response, remove_response: Response) -> Self {
-            Self {
-                remove_run_response: Some(remove_response),
-                ..Self::new(run_response)
-            }
-        }
-
-        fn with_frames(frames: Vec<Vec<u8>>) -> Self {
-            Self {
-                frames,
-                ..Self::new(Response::Pong)
-            }
+    fn running_run() -> lns_ipc::RunSummary {
+        lns_ipc::RunSummary {
+            id: "1a2b3c4d0000000000000000000000aa".into(),
+            name: "reviewer".into(),
+            image: "some-image".into(),
+            command: "cmd".into(),
+            status: lns_ipc::RunStatus::Running,
+            started: "2026-01-01T00:00:00Z".into(),
         }
     }
 
-    fn sandbox_inspection(tools: Vec<String>) -> Response {
-        sandbox_inspection_with_digest(tools, format!("sha256:{}", "a".repeat(64)))
-    }
-
-    fn sandbox_inspection_with_digest(tools: Vec<String>, digest: String) -> Response {
-        Response::ImageInspected {
-            inspection: lns_ipc::ArtifactInspection::Sandbox(Box::new(lns_ipc::SandboxView {
-                mixins: Vec::new(),
-                pinned_mixins: Vec::new(),
-                contributions: Vec::new(),
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                digest,
-                image: "docker.io/library/alpine@sha256:abc".into(),
-                workdir: None,
-                user: None,
-                mounts: Vec::new(),
-                ports: Vec::new(),
-                filesets: Vec::new(),
-                connectors: Vec::new(),
-                env: Vec::new(),
-                credentials: Vec::new(),
-                tools,
-                scripts: Vec::new(),
-                policy_flags: Vec::new(),
-                cpus: None,
-                mem_mib: None,
-                disk_bytes: None,
-            })),
+    fn running_inspect(status: lns_ipc::RunStatus) -> Response {
+        Response::RunInspect {
+            details: Box::new(RunDetails {
+                summary: lns_ipc::RunSummary {
+                    id: "1a2b3c4d0000000000000000000000aa".into(),
+                    name: "reviewer".into(),
+                    image: "some-image".into(),
+                    command: String::new(),
+                    status,
+                    started: "2026-01-01T00:00:00Z".into(),
+                },
+                config: lns_ipc::RunConfig::default(),
+            }),
         }
     }
 
-    fn pulled_response() -> Response {
-        Response::ImagePulled {
-            image: lns_ipc::ImageInfo {
-                kind: lns_ipc::CachedKind::Sandbox,
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                digest: format!("sha256:{}", "a".repeat(64)),
-                size_bytes: 1024,
-                layers: 1,
-                pulled: "2026-01-01T00:00:00Z".into(),
-                in_use_by: None,
+    fn ps_args() -> SandboxLsArgs {
+        SandboxLsArgs {
+            output: crate::output::OutputArgs {
+                format: crate::output::Format::Table,
             },
-            warnings: Vec::new(),
-        }
-    }
-
-    impl SandboxService for CannedService {
-        type Stream = tokio::io::DuplexStream;
-        fn one_shot(&self, request: Request) -> BoxFuture<'_, Result<Response>> {
-            let response = match &request {
-                Request::RunStats { .. } => self
-                    .stats_response
-                    .clone()
-                    .unwrap_or_else(|| self.response.clone()),
-                Request::InspectImage { .. } => self
-                    .inspect_image_response
-                    .clone()
-                    .unwrap_or_else(|| self.response.clone()),
-                Request::RemoveImage { .. } => self
-                    .remove_image_response
-                    .clone()
-                    .unwrap_or_else(|| self.response.clone()),
-                Request::RemoveRun { .. } => self
-                    .remove_run_response
-                    .clone()
-                    .unwrap_or_else(|| self.response.clone()),
-                _ => self.response.clone(),
-            };
-            self.requests.lock().unwrap().push(request);
-            Box::pin(async move { Ok(response) })
-        }
-        fn open_stream(&self, _request: Request) -> BoxFuture<'_, Result<Self::Stream>> {
-            let frames = self.frames.clone();
-            Box::pin(async move {
-                if frames.is_empty() {
-                    bail!("the daemon refused the stream");
-                }
-                Ok(stream_with(&frames).await)
-            })
-        }
-        fn aux_socket(&self) -> Option<PathBuf> {
-            None
-        }
-        fn load_policy(&self, _path: &str) -> Option<serde_json::Value> {
-            None
         }
     }
 
@@ -1633,7 +731,6 @@ mod tests {
             &cmd,
             &svc,
             TermInfo::default(),
-            &mut std::io::Cursor::new(""),
             &mut out,
             &mut stdout,
             &mut stderr,
@@ -1655,7 +752,6 @@ mod tests {
             &cmd,
             &svc,
             TermInfo::default(),
-            &mut std::io::Cursor::new(""),
             &mut out,
             &mut stdout,
             &mut stderr,
@@ -1663,452 +759,6 @@ mod tests {
         .await
         .unwrap_err();
         assert!(format!("{err:#}").contains("interactive"), "got: {err:#}");
-    }
-
-    #[tokio::test]
-    async fn run_with_writers_refuses_the_offline_author_verbs() {
-        let svc = CannedService::new(Response::Pong);
-        for cmd in [
-            SandboxCommand::Init(SandboxInitArgs {
-                kind: DocumentKind::Sandbox,
-                file: None,
-            }),
-            SandboxCommand::Validate(SandboxValidateArgs {
-                kind: None,
-                file: None,
-            }),
-            SandboxCommand::Inspect(SandboxInspectArgs {
-                run: None,
-                mixins: Vec::new(),
-                file: None,
-                output: crate::output::OutputArgs {
-                    format: crate::output::Format::Table,
-                },
-            }),
-        ] {
-            let mut out = Vec::new();
-            let mut stdout = Vec::new();
-            let mut stderr = Vec::new();
-            let err = run_with_writers(
-                &cmd,
-                &svc,
-                TermInfo::default(),
-                &mut std::io::Cursor::new(""),
-                &mut out,
-                &mut stdout,
-                &mut stderr,
-            )
-            .await
-            .unwrap_err();
-            assert!(format!("{err:#}").contains("offline"), "got: {err:#}");
-        }
-    }
-
-    #[tokio::test]
-    async fn run_with_writers_refuses_push_which_runs_locally() {
-        let svc = CannedService::new(Response::Pong);
-        let cmd = SandboxCommand::Push(SandboxPushArgs {
-            reference: "ghcr.io/team/hermes:1.4.0".into(),
-            dry_run: false,
-            assume_yes: false,
-            file: None,
-        });
-        let mut out = Vec::new();
-        let mut stdout = Vec::new();
-        let mut stderr = Vec::new();
-        let err = run_with_writers(
-            &cmd,
-            &svc,
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut out,
-            &mut stdout,
-            &mut stderr,
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("locally"), "got: {err:#}");
-    }
-
-    #[tokio::test]
-    async fn pull_reports_the_pulled_reference_and_digest() {
-        let digest = format!("sha256:{}", "a".repeat(64));
-        let svc = CannedService::with_inspect_image(
-            Response::ImagePulled {
-                image: lns_ipc::ImageInfo {
-                    kind: lns_ipc::CachedKind::Sandbox,
-                    reference: "ghcr.io/team/hermes:1.4.0".into(),
-                    digest: digest.clone(),
-                    size_bytes: 1024,
-                    layers: 1,
-                    pulled: "2026-01-01T00:00:00Z".into(),
-                    in_use_by: None,
-                },
-                warnings: vec!["the sandbox is cached, but its first run needs the network".into()],
-            },
-            sandbox_inspection(Vec::new()),
-        );
-        let mut out = Vec::new();
-        let mut stdout = Vec::new();
-        let mut stderr = Vec::new();
-        let code = run_with_writers(
-            &SandboxCommand::Pull(SandboxPullArgs {
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                assume_yes: false,
-            }),
-            &svc,
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut out,
-            &mut stdout,
-            &mut stderr,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        let text = String::from_utf8(out).unwrap();
-        assert!(text.contains("sha256:"), "got: {text}");
-        assert_eq!(
-            String::from_utf8(stderr).unwrap(),
-            "warning: the sandbox is cached, but its first run needs the network\n"
-        );
-        assert!(matches!(
-            svc.requests.lock().unwrap().as_slice(),
-            [
-                Request::InspectImage { image, .. },
-                Request::PullImage {
-                    image: pulled,
-                    expected_digest
-                }
-            ] if image == "ghcr.io/team/hermes:1.4.0"
-                && pulled == image
-                && expected_digest == &digest
-        ));
-    }
-
-    #[tokio::test]
-    async fn pull_discloses_declared_tools_before_requesting_provisioning() {
-        let svc = CannedService::with_inspect_image(
-            pulled_response(),
-            sandbox_inspection(vec!["node@22".into()]),
-        );
-        let mut out = Vec::new();
-
-        let code = pull(
-            &svc,
-            &SandboxPullArgs {
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                assume_yes: false,
-            },
-            TermInfo {
-                stdin_is_tty: true,
-                stdout_is_terminal: false,
-            },
-            &mut std::io::Cursor::new("yes\n"),
-            &mut out,
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(code, 0);
-        let text = String::from_utf8(out).unwrap();
-        assert!(text.contains("Tool:       node@22"), "got: {text}");
-        assert!(
-            matches!(
-                svc.requests.lock().unwrap().as_slice(),
-                [Request::InspectImage { .. }, Request::PullImage { .. }]
-            ),
-            "inspection must precede the provisioning pull"
-        );
-    }
-
-    #[tokio::test]
-    async fn declining_pulled_tools_sends_no_provisioning_request() {
-        let svc = CannedService::with_inspect_image(
-            pulled_response(),
-            sandbox_inspection(vec!["node@22".into()]),
-        );
-
-        let err = pull(
-            &svc,
-            &SandboxPullArgs {
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                assume_yes: false,
-            },
-            TermInfo {
-                stdin_is_tty: true,
-                stdout_is_terminal: false,
-            },
-            &mut std::io::Cursor::new("n\n"),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-
-        assert!(err.to_string().contains("declined"), "got: {err}");
-        assert!(matches!(
-            svc.requests.lock().unwrap().as_slice(),
-            [Request::InspectImage { .. }]
-        ));
-    }
-
-    #[tokio::test]
-    async fn noninteractive_tool_pull_requires_yes_and_yes_skips_the_prompt() {
-        let refused = CannedService::with_inspect_image(
-            pulled_response(),
-            sandbox_inspection(vec!["node@22".into()]),
-        );
-        let err = pull(
-            &refused,
-            &SandboxPullArgs {
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                assume_yes: false,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(err.to_string().contains("--yes"), "got: {err}");
-        assert!(matches!(
-            refused.requests.lock().unwrap().as_slice(),
-            [Request::InspectImage { .. }]
-        ));
-
-        let accepted = CannedService::with_inspect_image(
-            pulled_response(),
-            sandbox_inspection(vec!["node@22".into()]),
-        );
-        let mut out = Vec::new();
-        pull(
-            &accepted,
-            &SandboxPullArgs {
-                reference: "ghcr.io/team/hermes:1.4.0".into(),
-                assume_yes: true,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut out,
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap();
-        let out = String::from_utf8(out).unwrap();
-        assert!(!out.contains("Continue?"), "got: {out}");
-        assert!(matches!(
-            accepted.requests.lock().unwrap().as_slice(),
-            [Request::InspectImage { .. }, Request::PullImage { .. }]
-        ));
-    }
-
-    #[tokio::test]
-    async fn pull_surfaces_a_daemon_error_and_rejects_an_unrelated_variant() {
-        let err = pull(
-            &CannedService::new(Response::Error {
-                message: "registry unreachable".into(),
-            }),
-            &SandboxPullArgs {
-                reference: "x:1".into(),
-                assume_yes: false,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("registry unreachable"));
-
-        let err = pull(
-            &CannedService::new(Response::Pong),
-            &SandboxPullArgs {
-                reference: "x:1".into(),
-                assume_yes: false,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    #[tokio::test]
-    async fn pull_refuses_an_unpinned_sandbox_and_a_plain_image() {
-        let err = pull(
-            &CannedService::with_inspect_image(
-                Response::Pong,
-                sandbox_inspection_with_digest(Vec::new(), String::new()),
-            ),
-            &SandboxPullArgs {
-                reference: "x:1".into(),
-                assume_yes: false,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(err.to_string().contains("did not provide a digest"));
-
-        let err = pull(
-            &CannedService::with_inspect_image(
-                Response::Pong,
-                Response::ImageInspected {
-                    inspection: lns_ipc::ArtifactInspection::Image(lns_ipc::ImageView {
-                        reference: "x:1".into(),
-                        digest: format!("sha256:{}", "a".repeat(64)),
-                    }),
-                },
-            ),
-            &SandboxPullArgs {
-                reference: "x:1".into(),
-                assume_yes: false,
-            },
-            TermInfo::default(),
-            &mut std::io::Cursor::new(""),
-            &mut Vec::new(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(err.to_string().contains("OCI image"));
-    }
-
-    #[tokio::test]
-    async fn pull_surfaces_a_post_consent_daemon_error_and_unexpected_response() {
-        for (response, expected) in [
-            (
-                Response::Error {
-                    message: "registry changed".into(),
-                },
-                "registry changed",
-            ),
-            (Response::Pong, "unexpected response"),
-        ] {
-            let err = pull(
-                &CannedService::with_inspect_image(response, sandbox_inspection(Vec::new())),
-                &SandboxPullArgs {
-                    reference: "x:1".into(),
-                    assume_yes: false,
-                },
-                TermInfo::default(),
-                &mut std::io::Cursor::new(""),
-                &mut Vec::new(),
-                &mut Vec::new(),
-            )
-            .await
-            .unwrap_err();
-            assert!(err.to_string().contains(expected), "got: {err}");
-        }
-    }
-
-    #[tokio::test]
-    async fn tag_confirms_the_new_reference() {
-        let svc = CannedService::new(Response::ImageTagged {
-            from: "hermes:1.4.0".into(),
-            to: "hermes:latest".into(),
-        });
-        let mut out = Vec::new();
-        let code = tag(
-            &svc,
-            &SandboxTagArgs {
-                from: "hermes:1.4.0".into(),
-                to: "hermes:latest".into(),
-            },
-            &mut out,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        assert_eq!(
-            String::from_utf8(out).unwrap(),
-            "tagged hermes:1.4.0 as hermes:latest\n"
-        );
-    }
-
-    #[tokio::test]
-    async fn tag_surfaces_a_daemon_error_and_rejects_an_unrelated_variant() {
-        let err = tag(
-            &CannedService::new(Response::Error {
-                message: "no such cached sandbox".into(),
-            }),
-            &SandboxTagArgs {
-                from: "a:1".into(),
-                to: "a:2".into(),
-            },
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("no such cached sandbox"));
-
-        let err = tag(
-            &CannedService::new(Response::Pong),
-            &SandboxTagArgs {
-                from: "a:1".into(),
-                to: "a:2".into(),
-            },
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    #[tokio::test]
-    async fn ls_surfaces_a_daemon_error() {
-        let svc = CannedService::new(Response::Error {
-            message: "registry poisoned".into(),
-        });
-        let mut out = Vec::new();
-        let err = ls(&svc, &ls_args(), &mut out).await.unwrap_err();
-        assert!(format!("{err:#}").contains("registry poisoned"));
-    }
-
-    #[tokio::test]
-    async fn ls_rejects_an_unrelated_response_variant() {
-        let svc = CannedService::new(Response::Pong);
-        let mut out = Vec::new();
-        let err = ls(&svc, &ls_args(), &mut out).await.unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    fn ls_args() -> LsArgs {
-        LsArgs {
-            kind: None,
-            output: crate::output::OutputArgs {
-                format: crate::output::Format::Table,
-            },
-        }
-    }
-
-    fn ps_args() -> PsArgs {
-        PsArgs {
-            output: crate::output::OutputArgs {
-                format: crate::output::Format::Table,
-            },
-        }
-    }
-
-    fn running_run() -> lns_ipc::RunSummary {
-        lns_ipc::RunSummary {
-            id: "1a2b3c4d0000000000000000000000aa".into(),
-            name: "reviewer".into(),
-            image: "some-image".into(),
-            command: "cmd".into(),
-            status: lns_ipc::RunStatus::Running,
-            started: "2026-01-01T00:00:00Z".into(),
-        }
     }
 
     #[tokio::test]
@@ -2275,44 +925,10 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn inspect_of_a_live_run_refuses_a_mixin_rather_than_ignoring_it() {
-        let svc = CannedService::new(Response::RunInspect {
-            details: Box::new(RunDetails {
-                summary: lns_ipc::RunSummary {
-                    id: "1a2b3c4d0000000000000000000000aa".into(),
-                    name: "reviewer".into(),
-                    image: "some-image".into(),
-                    command: String::new(),
-                    status: lns_ipc::RunStatus::Running,
-                    started: "2026-01-01T00:00:00Z".into(),
-                },
-                config: lns_ipc::RunConfig {
-                    policy_path: Some("/work/lns-local-mixin.yaml".into()),
-                    ..Default::default()
-                },
-            }),
-        });
-        let mut out = Vec::new();
-        let err = inspect(
-            &svc,
-            "1",
-            &["ghcr.io/acme/obs:2".to_string()],
-            crate::output::Format::Json,
-            &mut out,
-        )
-        .await
-        .unwrap_err();
-        assert!(
-            format!("{err:#}").contains("--mixin applies to a sandbox reference"),
-            "a live run has already booted, so rendering it as though the flag applied would describe a composition that never ran; got: {err:#}"
-        );
-    }
-
-    #[tokio::test]
     async fn inspect_rejects_an_unrelated_response_variant() {
         let svc = CannedService::new(Response::Pong);
         let mut out = Vec::new();
-        let err = inspect(&svc, "1", &[], crate::output::Format::Json, &mut out)
+        let err = inspect(&svc, "1", crate::output::Format::Json, &mut out)
             .await
             .unwrap_err();
         assert!(format!("{err:#}").contains("unexpected response"));
@@ -2324,161 +940,10 @@ mod tests {
             message: "no active run with id 1".into(),
         });
         let mut out = Vec::new();
-        let err = inspect(&svc, "1", &[], crate::output::Format::Table, &mut out)
+        let err = inspect(&svc, "1", crate::output::Format::Table, &mut out)
             .await
             .unwrap_err();
         assert!(format!("{err:#}").contains("no active run with id 1"));
-    }
-
-    #[tokio::test]
-    async fn inspect_falls_back_to_a_cached_sandbox_definition() {
-        let svc = CannedService::with_inspect_image(
-            Response::Error {
-                message: "no such run: hermes:1.4.0".into(),
-            },
-            Response::ImageInspected {
-                inspection: lns_ipc::ArtifactInspection::Sandbox(Box::new(lns_ipc::SandboxView {
-                    mixins: Vec::new(),
-                    pinned_mixins: Vec::new(),
-                    contributions: Vec::new(),
-                    reference: "hermes:1.4.0".into(),
-                    digest: format!("sha256:{}", "a".repeat(64)),
-                    image: "docker.io/library/alpine@sha256:abc".into(),
-                    workdir: None,
-                    user: None,
-                    mounts: Vec::new(),
-                    ports: Vec::new(),
-                    filesets: Vec::new(),
-                    connectors: vec!["some-provider".into()],
-                    env: Vec::new(),
-                    credentials: Vec::new(),
-                    tools: vec!["node@22.11.0".into()],
-                    scripts: vec![
-                        lns_ipc::SandboxScript {
-                            when: "pre-start".into(),
-                            run: "apt-get update\napt-get install -y psql".into(),
-                            user: Some("root".into()),
-                            description: Some("the psql the prompts assume".into()),
-                        },
-                        lns_ipc::SandboxScript {
-                            when: "pre-start".into(),
-                            run: "npm ci".into(),
-                            user: None,
-                            description: None,
-                        },
-                    ],
-                    policy_flags: Vec::new(),
-                    cpus: None,
-                    mem_mib: None,
-                    disk_bytes: None,
-                })),
-            },
-        );
-        let mut out = Vec::new();
-        let code = inspect(
-            &svc,
-            "hermes:1.4.0",
-            &[],
-            crate::output::Format::Table,
-            &mut out,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        let text = String::from_utf8(out).unwrap();
-        assert!(text.contains("kind: sandbox"), "got: {text}");
-        assert!(
-            text.contains("image: docker.io/library/alpine"),
-            "got: {text}"
-        );
-        assert!(text.contains("connector: some-provider"), "got: {text}");
-        assert!(text.contains("tool: node@22.11.0"), "got: {text}");
-        assert!(
-            text.contains("script: pre-start (runs as root)")
-                && text.contains("  the psql the prompts assume")
-                && text.contains("  | apt-get update")
-                && text.contains("  | apt-get install -y psql"),
-            "inspect answers \"so what does it actually do\", so every line of the body has to appear; got: {text}"
-        );
-        assert!(
-            text.contains("script: pre-start (runs as the workload user)"),
-            "a script naming no user runs as the workload does, and a reader should not have to infer that from a blank; got: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn inspect_cached_renders_the_image_kind() {
-        let image = CannedService::with_inspect_image(
-            Response::Error {
-                message: "no such run: x".into(),
-            },
-            Response::ImageInspected {
-                inspection: lns_ipc::ArtifactInspection::Image(lns_ipc::ImageView {
-                    reference: "alpine:3.20".into(),
-                    digest: "sha256:abc".into(),
-                }),
-            },
-        );
-        let mut out = Vec::new();
-        inspect(&image, "x", &[], crate::output::Format::Table, &mut out)
-            .await
-            .unwrap();
-        let text = String::from_utf8(out).unwrap();
-        assert!(text.contains("kind: image"), "got: {text}");
-        assert!(text.contains("digest: sha256:abc"), "got: {text}");
-    }
-
-    #[tokio::test]
-    async fn inspect_cached_rejects_an_unrelated_image_response() {
-        let svc = CannedService::with_inspect_image(
-            Response::Error {
-                message: "no such run: x".into(),
-            },
-            Response::Pong,
-        );
-        let mut out = Vec::new();
-        let err = inspect(&svc, "x", &[], crate::output::Format::Table, &mut out)
-            .await
-            .unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    #[tokio::test]
-    async fn inspect_surfaces_a_non_not_found_inspect_error_without_touching_the_cache() {
-        let svc = CannedService::new(Response::Error {
-            message: "daemon busy".into(),
-        });
-        let mut out = Vec::new();
-        let err = inspect(&svc, "reviewer", &[], crate::output::Format::Json, &mut out)
-            .await
-            .unwrap_err();
-        assert!(
-            format!("{err:#}").contains("daemon busy"),
-            "a transient InspectRun error must surface, not be masked as no-such-image: {err:#}"
-        );
-        let requests = svc.requests.lock().unwrap();
-        assert!(
-            !requests
-                .iter()
-                .any(|r| matches!(r, Request::InspectImage { .. })),
-            "a genuine error must not fall through to the cached-inspect path: {requests:?}"
-        );
-    }
-
-    fn running_inspect(status: lns_ipc::RunStatus) -> Response {
-        Response::RunInspect {
-            details: Box::new(RunDetails {
-                summary: lns_ipc::RunSummary {
-                    id: "1a2b3c4d0000000000000000000000aa".into(),
-                    name: "reviewer".into(),
-                    image: "some-image".into(),
-                    command: String::new(),
-                    status,
-                    started: "2026-01-01T00:00:00Z".into(),
-                },
-                config: lns_ipc::RunConfig::default(),
-            }),
-        }
     }
 
     #[tokio::test]
@@ -2496,61 +961,6 @@ mod tests {
         .unwrap_err();
         assert!(format!("{err:#}").contains("running"), "got: {err:#}");
         assert!(format!("{err:#}").contains("lns stop"), "got: {err:#}");
-    }
-
-    #[tokio::test]
-    async fn rm_removes_a_cached_sandbox_and_reports_freed_layers() {
-        let svc = CannedService::with_remove_image(
-            Response::Error {
-                message: "no such run: hermes:1.4.0".into(),
-            },
-            Response::ImageRemoved {
-                reference: "hermes:1.4.0".into(),
-                reclaimed_bytes: 3 * 1024 * 1024,
-            },
-        );
-        let mut out = Vec::new();
-        let code = rm(
-            &svc,
-            &SandboxRmArgs {
-                run: "hermes:1.4.0".into(),
-            },
-            &mut out,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        let text = String::from_utf8(out).unwrap();
-        assert!(text.contains("removed hermes:1.4.0"), "got: {text}");
-        assert!(text.contains("freed 3.0 MiB"), "got: {text}");
-    }
-
-    #[tokio::test]
-    async fn rm_surfaces_a_non_not_found_inspect_error_without_touching_the_cache() {
-        let svc = CannedService::new(Response::Error {
-            message: "daemon busy".into(),
-        });
-        let mut out = Vec::new();
-        let err = rm(
-            &svc,
-            &SandboxRmArgs {
-                run: "reviewer".into(),
-            },
-            &mut out,
-        )
-        .await
-        .unwrap_err();
-        assert!(
-            format!("{err:#}").contains("daemon busy"),
-            "a transient InspectRun error must surface, not be masked as no-such-image: {err:#}"
-        );
-        let requests = svc.requests.lock().unwrap();
-        assert!(
-            !requests
-                .iter()
-                .any(|r| matches!(r, Request::RemoveImage { .. })),
-            "a genuine error must not fall through to the cached-remove path: {requests:?}"
-        );
     }
 
     #[tokio::test]
@@ -2635,145 +1045,11 @@ mod tests {
         assert!(format!("{err:#}").contains("unexpected response"));
     }
 
-    #[tokio::test]
-    async fn rm_of_a_cached_sandbox_surfaces_the_daemon_error() {
-        let svc = CannedService::with_remove_image(
-            Response::Error {
-                message: "no such run: ghcr.io/team/x:1".into(),
-            },
-            Response::Error {
-                message: "no such image: ghcr.io/team/x:1".into(),
-            },
-        );
-        let mut out = Vec::new();
-        let err = rm(
-            &svc,
-            &SandboxRmArgs {
-                run: "ghcr.io/team/x:1".into(),
-            },
-            &mut out,
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("no such image"));
-    }
-
-    #[tokio::test]
-    async fn rm_rejects_an_unrelated_remove_response() {
-        let svc = CannedService::with_remove_image(
-            Response::Error {
-                message: "no such run: reviewer".into(),
-            },
-            Response::Pong,
-        );
-        let mut out = Vec::new();
-        let err = rm(
-            &svc,
-            &SandboxRmArgs {
-                run: "reviewer".into(),
-            },
-            &mut out,
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    #[tokio::test]
-    async fn declining_the_prune_prompt_reaches_no_service() {
-        let svc = CannedService::new(Response::Pong);
-        let mut out = Vec::new();
-        let code = prune(
-            &svc,
-            &SandboxPruneArgs { force: false },
-            TermInfo {
-                stdin_is_tty: true,
-                stdout_is_terminal: false,
-            },
-            &mut std::io::Cursor::new(b"n\n".to_vec()),
-            &mut out,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        assert!(String::from_utf8(out).unwrap().contains("Aborted."));
-        assert!(svc.requests.lock().unwrap().is_empty());
-    }
-
-    #[tokio::test]
-    async fn prune_with_force_lists_removed_sandboxes_and_reclaimed_bytes() {
-        let svc = CannedService::new(Response::ImagesPruned {
-            removed: vec!["b:2".into(), "a:1".into()],
-            reclaimed_bytes: 64 * 1024 * 1024,
-        });
-        let mut out = Vec::new();
-        let code = prune(
-            &svc,
-            &SandboxPruneArgs { force: true },
-            TermInfo::default(),
-            &mut std::io::empty(),
-            &mut out,
-        )
-        .await
-        .unwrap();
-        assert_eq!(code, 0);
-        let text = String::from_utf8(out).unwrap();
-        assert!(
-            text.contains("removed a:1") && text.contains("removed b:2"),
-            "got: {text}"
-        );
-        assert!(text.contains("reclaimed 64.0 MiB"), "got: {text}");
-    }
-
-    #[tokio::test]
-    async fn prune_surfaces_a_daemon_error_and_rejects_an_unrelated_variant() {
-        let err = prune(
-            &CannedService::new(Response::Error {
-                message: "registry poisoned".into(),
-            }),
-            &SandboxPruneArgs { force: true },
-            TermInfo::default(),
-            &mut std::io::empty(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("registry poisoned"));
-
-        let err = prune(
-            &CannedService::new(Response::Pong),
-            &SandboxPruneArgs { force: true },
-            TermInfo::default(),
-            &mut std::io::empty(),
-            &mut Vec::new(),
-        )
-        .await
-        .unwrap_err();
-        assert!(format!("{err:#}").contains("unexpected response"));
-    }
-
-    #[test]
-    fn format_bytes_picks_the_natural_unit() {
-        assert_eq!(format_bytes(512), "512 B");
-        assert_eq!(format_bytes(2048), "2.0 KiB");
-        assert_eq!(format_bytes(92_274_688), "88.0 MiB");
-        assert_eq!(format_bytes(2 * 1024 * 1024 * 1024), "2.0 GiB");
-    }
-
     #[test]
     fn format_permille_renders_tenths_of_a_percent() {
         assert_eq!(format_permille(0), "0.0%");
         assert_eq!(format_permille(125), "12.5%");
         assert_eq!(format_permille(1000), "100.0%");
-    }
-
-    async fn stream_with(frames: &[Vec<u8>]) -> tokio::io::DuplexStream {
-        let (client, mut server) = tokio::io::duplex(64 * 1024);
-        let payload: Vec<u8> = frames.concat();
-        tokio::spawn(async move {
-            let _ = server.write_all(&payload).await;
-        });
-        client
     }
 
     #[tokio::test]
@@ -2896,7 +1172,7 @@ mod tests {
             }),
         });
         let mut out = Vec::new();
-        let code = inspect(&svc, "1", &[], crate::output::Format::Json, &mut out)
+        let code = inspect(&svc, "1", crate::output::Format::Json, &mut out)
             .await
             .unwrap();
         assert_eq!(code, 0);
