+++ crates/lns-cli/src/service/client.rs
@@ -1,8 +1,10 @@
 use anyhow::Result;
-use lns_ipc::StatusInfo;
+use lns_ipc::{Request, Response, StatusInfo};
 use std::future::Future;
+use std::path::PathBuf;
 use std::pin::Pin;
 use std::time::Duration;
+use tokio::io::{AsyncRead, AsyncWrite};
 
 pub type BoxFuture<'a, T> = Pin<Box<dyn Future<Output = T> + Send + 'a>>;
 
@@ -14,4 +16,19 @@ pub trait ServiceClient: Send + Sync {
     fn wait_for_ready(&self, total_timeout: Duration) -> BoxFuture<'_, bool>;
     fn wait_for_stopped(&self, total_timeout: Duration) -> BoxFuture<'_, bool>;
     fn cancel_run(&self, run_id: String) -> BoxFuture<'_, ()>;
+}
+
+/// The one request surface both namespaces drive the service through: `lns artifact` for what the cache holds, `lns sandbox` for what is running.
+pub trait SandboxService: Send + Sync {
+    type Stream: AsyncRead + AsyncWrite + Unpin + Send + 'static;
+    fn one_shot(&self, request: Request) -> BoxFuture<'_, Result<Response>>;
+    fn open_stream(&self, request: Request) -> BoxFuture<'_, Result<Self::Stream>>;
+    fn aux_socket(&self) -> Option<PathBuf>;
+    fn load_policy(&self, path: &str) -> Option<serde_json::Value>;
+}
+
+#[derive(Debug, Clone, Copy, Default)]
+pub struct TermInfo {
+    pub stdin_is_tty: bool,
+    pub stdout_is_terminal: bool,
 }
