+++ crates/lns-cli/tests/behaviours/machine_output.feature
@@ -6,7 +6,7 @@ Feature: machine-readable output from the list verbs
 
   Scenario: ps emits one JSON object per running sandbox
     Given the service reports one running sandbox using 125 permille cpu and 92274688 bytes
-    When the user runs sandbox command "ps --format json"
+    When the user runs sandbox command "ls --format json"
     Then the exit code is 0
     And the output is a JSON array of 1 rows
     And JSON row 0 has "name" set to "reviewer"
@@ -17,14 +17,14 @@ Feature: machine-readable output from the list verbs
 
   Scenario: ps json carries the full run id the table abbreviates
     Given the service reports one running sandbox using 125 permille cpu and 92274688 bytes
-    When the user runs sandbox command "ps --format json"
+    When the user runs sandbox command "ls --format json"
     Then the exit code is 0
     And JSON row 0 has "id" set to "00000003000000000000000000000000"
     And the output does not contain "000000030000  "
 
   Scenario: ps nulls the stats of a run whose guest stopped answering
     Given the service reports one running sandbox whose stats probe fails
-    When the user runs sandbox command "ps --format json"
+    When the user runs sandbox command "ls --format json"
     Then the exit code is 0
     And the output is a JSON array of 1 rows
     And JSON row 0 has "name" set to "reviewer"
@@ -34,27 +34,27 @@ Feature: machine-readable output from the list verbs
 
   Scenario: the human table keeps a run whose guest stopped answering
     Given the service reports one running sandbox whose stats probe fails
-    When the user runs sandbox command "ps"
+    When the user runs sandbox command "ls"
     Then the exit code is 0
     And the output contains "reviewer"
     And the output does not contain "sampling guest stats failed"
 
   Scenario: ps with nothing running is an empty array, not absent output
     Given the service reports no runs
-    When the user runs sandbox command "ps --format json"
+    When the user runs sandbox command "ls --format json"
     Then the exit code is 0
     And the output is an empty JSON array
 
   Scenario: the human table stays the default
     Given the service reports one running sandbox using 125 permille cpu and 92274688 bytes
-    When the user runs sandbox command "ps"
+    When the user runs sandbox command "ls"
     Then the exit code is 0
     And the output contains "CPU %"
     And the output contains "88.0 MiB"
 
-  Scenario: sandbox ls json exposes the fields the table has no room for
+  Scenario: artifact ls json exposes the fields the table has no room for
     Given the service reports one cached sandbox "hermes:1.4.0"
-    When the user runs sandbox command "ls --format json"
+    When the user runs artifact command "ls --format json"
     Then the exit code is 0
     And the output is a JSON array of 1 rows
     And JSON row 0 has "reference" set to "hermes:1.4.0"
@@ -64,9 +64,9 @@ Feature: machine-readable output from the list verbs
     And JSON row 0 has "pulled" set to "2026-01-01T00:00:00Z"
     And JSON row 0 has a null "inUseBy"
 
-  Scenario: sandbox ls with an empty cache is an empty array
+  Scenario: artifact ls with an empty cache is an empty array
     Given the service reports no cached sandboxes
-    When the user runs sandbox command "ls --format json"
+    When the user runs artifact command "ls --format json"
     Then the exit code is 0
     And the output is an empty JSON array
