+++ crates/lns-cli/tests/behaviours/sandbox_inspect_cli.feature
@@ -56,7 +56,7 @@ Feature: inspecting a typed artifact before running it
 
   Scenario: inspect discloses the run-as user a pulled sandbox asks for
     Given the service inspects "registry.example.test/some-sandbox:1.0" as a sandbox declaring user "root"
-    When the user runs sandbox command "inspect registry.example.test/some-sandbox:1.0"
+    When the user runs artifact command "inspect registry.example.test/some-sandbox:1.0"
     Then the exit code is 0
     And the output contains "user: root"
 
@@ -92,6 +92,6 @@ Feature: inspecting a typed artifact before running it
 
   Scenario: inspecting names the host that needs a login
     Given the service reports "inspect" needs a login for host "other-registry.example.test"
-    When the user runs "lns inspect some-registry.example/some-sandbox:research"
+    When the user runs artifact command "inspect some-registry.example/some-sandbox:research"
     Then the exit code is 1
     And the output contains "other-registry.example.test"
